requirements-dev.txt
bedrock-agentcore==0.1.7
bedrock-agentcore-starter-toolkit==0.1.19
strands-agents==1.10.0
strands-agents-tools==0.2.9
strands-agents-builder==0.1.10
langgraph==0.6.8
langchain-community==0.3.30
langchain-aws==0.2.35
langchain-ollama==0.3.10
mcp>=1.10.0
langchain-postgres==0.0.15
langchain-core==0.3.78
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
boto3==1.40.45
tavily-python==0.7.12
grandalf==0.8
